Fix h2 names and update assignment lists
Cleaned up header names/titles to make more sense. Linked to the
rest of the practice assignments.

Trying to establish a consistent layout/theme using the few HTML
tools I've learned so far. Need to layout a clear folder hierarchy
right on the main page at somepoint. Not a huge fan of the current
category and lists structure at the moment.